backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
starlette==0.46.2
pydantic==2.11.3
SQLAlchemy==2.0.40
alembic == 1.15.2
jinja2==3.1.4
python-jose==3.4.0
python-multipart==0.0.20
passlib==1.7.4
email-validator==2.2.0
mysqlclient==2.2.7
bcrypt==4.3.0
